dbShop/requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
dnspython==2.2.1
drf-yasg==1.21.4
eventlet==0.33.2
flower==1.2.0
greenlet==2.0.1
gunicorn==20.1.0
humanize==4.4.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
packaging==21.3
prometheus-client==0.15.0
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
PyJWT==2.6.0
pyparsing==3.0.9
pytz==2022.6
redis==4.4.2
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
tornado==6.2
tzdata==2022.6
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
